+++ NEWS.md
@@ -1,5 +1,6 @@
 # ggtibble 1.0.0.9000
 
+* `labs` argument to `ggtibble()` can now include `NULL (#6)
 * `guides()` can now be added to `gglist` objects.
 * Labels created with the `labs` argument to `ggtibble()` will not longer all be
   the same (#3)